make sure one can resize the header of the parametertree
import from the subclassed parameter tree and stuff to ovoid circular imports
