+++ src/pymodaq/control_modules/daq_viewer_ui.py
@@ -356,7 +356,7 @@ class DAQ_Viewer_UI(ControlModuleUI, ViewerDispatcher):
 
 def main(init_qt=True):
     from pymodaq.utils.gui_utils.dock import DockArea
-    from pymodaq.utils.managers.parameter_manager import ParameterTree, Parameter
+    from pymodaq.utils.parameter import ParameterTree, Parameter
     from pymodaq.control_modules.viewer_utility_classes import params as daq_viewer_params
 
     if init_qt:  # used for the test suite
