+++ src/pymodaq/control_modules/utils.py
@@ -291,7 +291,8 @@ class ControlModule(QObject):
     def show_config(self, config: Config):
         """ Display in a tree the current configuration"""
         if config is not None:
-            config_tree = configmod.TreeFromToml(config)
+            from pymodaq.utils.gui_utils.widgets.tree_toml import TreeFromToml
+            config_tree = TreeFromToml(config)
             config_tree.show_dialog()
 
     def update_status(self, txt, log=True):
