+++ src/pymodaq/dashboard.py
@@ -8,7 +8,6 @@ import logging
 from pathlib import Path
 from importlib import import_module
 from packaging import version as version_mod
-from pyqtgraph.parametertree import Parameter, ParameterTree
 from qtpy import QtGui, QtWidgets, QtCore
 from qtpy.QtCore import Qt, QObject, Slot, QThread, Signal
 from time import perf_counter
@@ -28,6 +27,7 @@ from pymodaq.utils.managers.remote_manager import RemoteManager
 from pymodaq.utils.managers.roi_manager import ROISaver
 from pymodaq.utils.exceptions import DetectorError, ActuatorError
 from pymodaq.utils import config as configmod
+from pymodaq.utils.parameter import ParameterTree, Parameter
 
 from pymodaq.control_modules.daq_move import DAQ_Move
 from pymodaq.control_modules.daq_viewer import DAQ_Viewer
@@ -1192,7 +1192,8 @@ class DashBoard(QObject):
         webbrowser.open(logging.getLogger('pymodaq').handlers[0].baseFilename)
 
     def show_config(self):
-        config_tree = configmod.TreeFromToml()
+        from pymodaq.utils.gui_utils.widgets.tree_toml import TreeFromToml
+        config_tree = TreeFromToml()
         config_tree.show_dialog()
 
     def setupUI(self):
